refactor: break down core.rs into modular architecture
