--- crates/cache/src/core.rs
@@ -1,2068 +1,0 @@
-//! Production-ready unified cache implementation
-//!
-//! This module provides a Google-scale cache implementation with:
-//! - Zero-copy architecture using memory-mapped files
-//! - 4-level sharding for optimal file system performance
-//! - Separate metadata storage for efficient scanning
-//! - No `?` operators - explicit error handling only
-//! - Lock-free concurrent access patterns
-//! - Comprehensive observability and metrics
-
-pub mod internal;
-pub mod serialization;
-
-use crate::cleanup::CacheCleanup;
-use self::internal::{CacheStats, InMemoryEntry, MemoryMapUtils, PathUtils};
-use self::serialization::{deserialize, serialize};
-use crate::errors::{CacheError, RecoveryHint, Result, SerializationOp, StoreType};
-use crate::eviction::{create_eviction_policy, EvictionPolicy};
-use crate::fast_path::FastPathCache;
-use crate::memory_manager::{MemoryManager, MemoryThresholds};
-use crate::streaming::{CacheReader, CacheWriter, StreamingCache};
-use crate::traits::{Cache as CacheTrait, CacheConfig, CacheKey, CacheMetadata, CacheStatistics};
-use async_trait::async_trait;
-use dashmap::DashMap;
-use futures::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
-use memmap2::{Mmap, MmapOptions};
-use parking_lot::RwLock;
-use serde::{de::DeserializeOwned, Serialize};
-use sha2::{Digest, Sha256};
-use std::fmt;
-use std::fs::File;
-use std::future::Future;
-use std::path::PathBuf;
-use std::pin::Pin;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::Arc;
-use std::time::{Duration, Instant, SystemTime};
-use tokio::fs;
-use tokio::sync::Semaphore;
-use tokio::task::JoinHandle;
-
-/// Production-ready unified cache implementation
-#[derive(Clone)]
-pub struct Cache {
-    inner: Arc<CacheInner>,
-}
-
-struct CacheInner {
-    /// Configuration
-    config: CacheConfig,
-    /// Base directory for file-based cache
-    base_dir: PathBuf,
-    /// In-memory cache for hot data
-    memory_cache: DashMap<String, Arc<InMemoryEntry>>,
-    /// Fast path cache for small values
-    fast_path: FastPathCache,
-    /// Eviction policy
-    eviction_policy: Box<dyn EvictionPolicy>,
-    /// Memory manager
-    memory_manager: Arc<MemoryManager>,
-    /// Statistics
-    stats: CacheStats,
-    /// Semaphore for limiting concurrent read operations
-    read_semaphore: Semaphore,
-    /// Semaphore for limiting concurrent write operations
-    write_semaphore: Semaphore,
-    /// Background cleanup task handle
-    cleanup_handle: RwLock<Option<JoinHandle<()>>>,
-    /// Cache format version
-    version: u32,
-}
-
-
-impl Cache {
-    /// Create a new unified cache with production-ready features
-    pub async fn new(base_dir: PathBuf, config: CacheConfig) -> Result<Self> {
-        // Create cache directories
-        match fs::create_dir_all(&base_dir).await {
-            Ok(()) => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: base_dir.clone(),
-                    operation: "create cache directory",
-                    source: e,
-                    recovery_hint: RecoveryHint::CheckPermissions {
-                        path: base_dir.clone(),
-                    },
-                });
-            }
-        }
-
-        // Create objects directory with 4-level sharding
-        let objects_dir = base_dir.join("objects");
-        match fs::create_dir_all(&objects_dir).await {
-            Ok(()) => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: objects_dir.clone(),
-                    operation: "create objects directory",
-                    source: e,
-                    recovery_hint: RecoveryHint::CheckPermissions { path: objects_dir },
-                });
-            }
-        }
-
-        // Create metadata directory structure
-        let metadata_dir = base_dir.join("metadata");
-        match fs::create_dir_all(&metadata_dir).await {
-            Ok(()) => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: metadata_dir.clone(),
-                    operation: "create metadata directory",
-                    source: e,
-                    recovery_hint: RecoveryHint::CheckPermissions { path: metadata_dir },
-                });
-            }
-        }
-
-        // Create eviction policy
-        let eviction_policy = match create_eviction_policy(
-            config.eviction_policy.as_deref().unwrap_or("lru"),
-            config.max_memory_size.unwrap_or(1024 * 1024 * 1024), // 1GB default
-        ) {
-            Ok(policy) => policy,
-            Err(e) => return Err(e),
-        };
-
-        // Create memory manager
-        let memory_manager = Arc::new(MemoryManager::new(
-            base_dir.clone(),
-            config.max_disk_size.unwrap_or(10 * 1024 * 1024 * 1024), // 10GB default
-            MemoryThresholds::default(),
-        ));
-
-        // Start memory monitoring
-        let manager_clone = Arc::clone(&memory_manager);
-        manager_clone.start_monitoring();
-
-        let inner = Arc::new(CacheInner {
-            config,
-            base_dir,
-            memory_cache: DashMap::new(),
-            fast_path: FastPathCache::new(1024, 10000), // 1KB threshold, 10k entries
-            eviction_policy,
-            memory_manager,
-            stats: CacheStats {
-                hits: AtomicU64::new(0),
-                misses: AtomicU64::new(0),
-                writes: AtomicU64::new(0),
-                removals: AtomicU64::new(0),
-                errors: AtomicU64::new(0),
-                total_bytes: AtomicU64::new(0),
-                expired_cleanups: AtomicU64::new(0),
-                entry_count: AtomicU64::new(0),
-                stats_since: SystemTime::now(),
-            },
-            read_semaphore: Semaphore::new(200), // More permits for reads
-            write_semaphore: Semaphore::new(50), // Fewer permits for writes
-            cleanup_handle: RwLock::new(None),
-            version: 3, // Version 3 with streaming and performance optimizations
-        });
-
-        let cache = Self { inner };
-
-        // Start background cleanup task
-        cache.start_cleanup_task();
-
-        Ok(cache)
-    }
-
-    /// Start the background cleanup task
-    fn start_cleanup_task(&self) {
-        let inner = Arc::clone(&self.inner);
-        let cleanup_interval = inner.config.cleanup_interval;
-
-        // Don't start cleanup task if interval is zero (useful for tests)
-        if cleanup_interval == Duration::ZERO {
-            return;
-        }
-
-        let handle = tokio::spawn(async move {
-            let mut interval = tokio::time::interval(cleanup_interval);
-            interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-
-            loop {
-                interval.tick().await;
-                match Self::cleanup_expired_entries(&inner).await {
-                    Ok(()) => {}
-                    Err(e) => {
-                        tracing::warn!("Cache cleanup error: {}", e);
-                    }
-                }
-            }
-        });
-
-        *self.inner.cleanup_handle.write() = Some(handle);
-    }
-
-    /// Clean up expired entries and corrupted files
-    async fn cleanup_expired_entries(inner: &Arc<CacheInner>) -> Result<()> {
-        let now = SystemTime::now();
-        let mut expired_keys = Vec::new();
-        let mut corrupted_keys = Vec::new();
-
-        // Find expired entries in memory
-        for entry in inner.memory_cache.iter() {
-            if let Some(expires_at) = entry.value().metadata.expires_at {
-                if expires_at <= now {
-                    expired_keys.push(entry.key().clone());
-                }
-            }
-        }
-
-        // Scan disk for orphaned metadata files (no corresponding data file)
-        let metadata_dir = inner.base_dir.join("metadata");
-        if fs::metadata(&metadata_dir).await.is_ok() {
-            Self::scan_for_corrupted_files(inner, &metadata_dir, &mut corrupted_keys).await;
-        }
-
-        // Remove expired entries
-        for key in expired_keys {
-            if let Some((_, entry)) = inner.memory_cache.remove(&key) {
-                inner
-                    .stats
-                    .total_bytes
-                    .fetch_sub(entry.data.len() as u64, Ordering::Relaxed);
-                inner.stats.entry_count.fetch_sub(1, Ordering::Relaxed);
-                inner.stats.expired_cleanups.fetch_add(1, Ordering::Relaxed);
-            }
-
-            Self::cleanup_entry_files(inner, &key).await;
-        }
-
-        // Remove corrupted entries
-        for key in corrupted_keys {
-            // Remove from memory cache if present
-            inner.memory_cache.remove(&key);
-            Self::cleanup_entry_files(inner, &key).await;
-            inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-        }
-
-        Ok(())
-    }
-
-    /// Scan for corrupted files in the top level only (non-recursive for now)
-    async fn scan_for_corrupted_files(
-        inner: &Arc<CacheInner>,
-        metadata_dir: &std::path::Path,
-        _corrupted_keys: &mut [String],
-    ) {
-        // Simple non-recursive cleanup to avoid boxing issues
-        // Just clean up obvious orphaned files in the metadata directory
-        let mut read_dir = match tokio::fs::read_dir(metadata_dir).await {
-            Ok(dir) => dir,
-            Err(_) => return,
-        };
-
-        let mut cleanup_count = 0;
-        while let Ok(Some(entry)) = read_dir.next_entry().await {
-            let path = entry.path();
-            if !path.is_dir() && path.extension().and_then(|s| s.to_str()) == Some("meta") {
-                // Check if corresponding data file exists
-                if let Some(file_stem) = path.file_stem().and_then(|s| s.to_str()) {
-                    let data_path = Self::object_path_from_hash(inner, file_stem);
-                    let data_exists = fs::metadata(&data_path).await.is_ok();
-                    if !data_exists {
-                        // Orphaned metadata file - clean it up
-                        if fs::remove_file(&path).await.is_ok() {
-                            cleanup_count += 1;
-                            tracing::debug!(
-                                "Cleaned up orphaned metadata file: {}",
-                                path.display()
-                            );
-                        }
-                    }
-                }
-            }
-
-            // Limit cleanup per run to avoid blocking too long
-            if cleanup_count >= 50 {
-                break;
-            }
-        }
-    }
-
-    /// Get object path from hash (for cleanup)
-    fn object_path_from_hash(inner: &CacheInner, hash: &str) -> PathBuf {
-        let shard = &hash[..2];
-        inner.base_dir.join("objects").join(shard).join(hash)
-    }
-
-    /// Clean up files for a specific entry
-    async fn cleanup_entry_files(inner: &Arc<CacheInner>, key: &str) {
-        let metadata_path = Self::metadata_path(inner, key);
-        let data_path = Self::object_path(inner, key);
-
-        match fs::remove_file(&metadata_path).await {
-            Ok(()) => {}
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-            Err(e) => {
-                tracing::warn!(
-                    "Failed to remove expired metadata {}: {}",
-                    metadata_path.display(),
-                    e
-                );
-            }
-        }
-
-        match fs::remove_file(&data_path).await {
-            Ok(()) => {}
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-            Err(e) => {
-                tracing::warn!(
-                    "Failed to remove expired data {}: {}",
-                    data_path.display(),
-                    e
-                );
-            }
-        }
-    }
-
-    /// Get the path for a cached object using optimized 256-shard distribution
-    #[inline(always)]
-    fn object_path(inner: &CacheInner, key: &str) -> PathBuf {
-        let hash = Self::hash_key(inner, key);
-        // Use first byte of hash for 256-way sharding (00-ff)
-        // This provides optimal distribution for file systems
-        let shard = &hash[..2];
-        inner.base_dir.join("objects").join(shard).join(&hash)
-    }
-
-    /// Get the path for cached metadata using optimized 256-shard distribution
-    #[inline(always)]
-    fn metadata_path(inner: &CacheInner, key: &str) -> PathBuf {
-        let hash = Self::hash_key(inner, key);
-        // Use first byte of hash for 256-way sharding (00-ff)
-        let shard = &hash[..2];
-        inner
-            .base_dir
-            .join("metadata")
-            .join(shard)
-            .join(format!("{}.meta", &hash))
-    }
-
-    /// Hash a cache key with performance optimizations
-    #[inline(always)]
-    fn hash_key(inner: &CacheInner, key: &str) -> String {
-        // Use SIMD-accelerated hashing when available
-        #[cfg(target_arch = "x86_64")]
-        {
-            use crate::performance::simd_hash;
-            if simd_hash::is_simd_available() {
-                let simd_hash = unsafe { simd_hash::hash_key_simd(key.as_bytes()) };
-                // Mix with version for cache invalidation
-                let mixed = simd_hash ^ (inner.version as u64);
-                return format!("{mixed:016x}");
-            }
-        }
-
-        // Fallback to SHA256 for cryptographic strength
-        let mut hasher = Sha256::new();
-        hasher.update(key.as_bytes());
-        hasher.update(inner.version.to_le_bytes());
-        format!("{:x}", hasher.finalize())
-    }
-
-    /// Serialize a value
-    fn serialize<T: Serialize>(value: &T) -> Result<Vec<u8>> {
-        match bincode::serialize(value) {
-            Ok(bytes) => Ok(bytes),
-            Err(e) => Err(CacheError::Serialization {
-                key: String::new(),
-                operation: SerializationOp::Encode,
-                source: Box::new(e),
-                recovery_hint: RecoveryHint::Manual {
-                    instructions: "Check that the value is serializable".to_string(),
-                },
-            }),
-        }
-    }
-
-    /// Deserialize a value
-    fn deserialize<T: DeserializeOwned>(data: &[u8]) -> Result<T> {
-        match bincode::deserialize(data) {
-            Ok(value) => Ok(value),
-            Err(e) => Err(CacheError::Serialization {
-                key: String::new(),
-                operation: SerializationOp::Decode,
-                source: Box::new(e),
-                recovery_hint: RecoveryHint::ClearAndRetry,
-            }),
-        }
-    }
-
-    /// Memory-map a file for zero-copy access
-    fn mmap_file(path: &PathBuf) -> Result<Mmap> {
-        let file = match File::open(path) {
-            Ok(f) => f,
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: path.clone(),
-                    operation: "open file for mmap",
-                    source: e,
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(10),
-                    },
-                });
-            }
-        };
-
-        match unsafe { MmapOptions::new().map(&file) } {
-            Ok(mmap) => Ok(mmap),
-            Err(e) => Err(CacheError::Io {
-                path: path.clone(),
-                operation: "memory-map file",
-                source: e,
-                recovery_hint: RecoveryHint::Manual {
-                    instructions: "Check available memory and file permissions".to_string(),
-                },
-            }),
-        }
-    }
-
-    /// Evict entries based on the configured eviction policy
-    async fn evict_entries(&self) -> Result<()> {
-        let mut evicted_count = 0;
-
-        // Use cache's actual memory usage, not the eviction policy's tracking
-        let max_memory = match self.inner.config.max_memory_size {
-            Some(max) => max,
-            None => return Ok(()), // No memory limit configured
-        };
-
-        let target_memory = max_memory * 8 / 10; // Target 80% usage
-
-        loop {
-            // Check current cache memory usage
-            let current_memory = self.inner.stats.total_bytes.load(Ordering::Relaxed);
-
-            // Check if we need to evict more
-            if current_memory <= target_memory {
-                break;
-            }
-
-            // Get next key to evict
-            let mut key_to_evict = self.inner.eviction_policy.next_eviction();
-
-            // Fallback: if the policy can't provide a key (e.g., under contention),
-            // choose the least-recently-accessed in-memory entry to ensure forward progress.
-            if key_to_evict.is_none() {
-                let mut oldest_key: Option<String> = None;
-                let mut oldest_instant = Instant::now();
-                for item in self.inner.memory_cache.iter() {
-                    let last = *item.value().last_accessed.read();
-                    if oldest_key.is_none() || last < oldest_instant {
-                        oldest_instant = last;
-                        oldest_key = Some(item.key().clone());
-                    }
-                }
-                key_to_evict = oldest_key;
-            }
-
-            let key_to_evict = match key_to_evict {
-                Some(key) => key,
-                None => break, // No more entries to evict
-            };
-
-            // Remove the entry
-            match self.remove(&key_to_evict).await {
-                Ok(removed) => {
-                    if removed {
-                        evicted_count += 1;
-                    }
-                }
-                Err(e) => {
-                    tracing::warn!("Failed to evict {}: {}", key_to_evict, e);
-                }
-            }
-
-            // Limit evictions per call to prevent blocking too long
-            if evicted_count >= 100 {
-                break;
-            }
-        }
-
-        if evicted_count > 0 {
-            tracing::info!("Evicted {} entries to free memory", evicted_count);
-        }
-
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl CacheTrait for Cache {
-    async fn get<T>(&self, key: &str) -> Result<Option<T>>
-    where
-        T: DeserializeOwned + Send + 'static,
-    {
-        match key.validate() {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        // Check fast-path cache for small values
-        if let Some((data, _metadata)) = self.inner.fast_path.get_small(key) {
-            // Deserialize the data from fast path
-            match Self::deserialize::<T>(&data) {
-                Ok(value) => {
-                    self.inner.stats.hits.fetch_add(1, Ordering::Relaxed);
-                    return Ok(Some(value));
-                }
-                Err(_) => {
-                    // If deserialization fails, remove from fast path and continue to regular path
-                    self.inner.fast_path.remove_small(key);
-                }
-            }
-        }
-
-        // Check memory cache first
-        if let Some(entry) = self.inner.memory_cache.get(key) {
-            // Update access time
-            *entry.last_accessed.write() = Instant::now();
-
-            // Check if expired
-            if let Some(expires_at) = entry.metadata.expires_at {
-                if expires_at <= SystemTime::now() {
-                    // Remove expired entry
-                    drop(entry);
-                    self.inner.memory_cache.remove(key);
-                    self.inner.stats.misses.fetch_add(1, Ordering::Relaxed);
-                    return Ok(None);
-                }
-            }
-
-            // Record access for eviction policy
-            let size = if entry.mmap.is_some() {
-                entry.metadata.size_bytes
-            } else {
-                entry.data.len() as u64
-            };
-            self.inner.eviction_policy.on_access(key, size);
-
-            self.inner.stats.hits.fetch_add(1, Ordering::Relaxed);
-
-            // Use memory-mapped data if available
-            let data = if let Some(ref mmap) = entry.mmap {
-                &mmap.as_ref()[..]
-            } else {
-                &entry.data
-            };
-
-            {
-                // Special-case: if caller expects Vec<u8> but data was encoded as Vec<i32> (due to type inference on put),
-                // detect and convert: [len:u64][i32 LE]* -> Vec<u8> by truncating each element.
-                use std::any::{Any, TypeId};
-                if TypeId::of::<T>() == TypeId::of::<Vec<u8>>() && data.len() >= 8 {
-                    let mut len_bytes = [0u8; 8];
-                    len_bytes.copy_from_slice(&data[..8]);
-                    let n = u64::from_le_bytes(len_bytes);
-                    let expected_len = 8usize.saturating_add((n as usize).saturating_mul(4));
-                    if expected_len == data.len() {
-                        let mut out = Vec::with_capacity(n as usize);
-                        for i in 0..(n as usize) {
-                            let start = 8 + i * 4;
-                            let mut w = [0u8; 4];
-                            w.copy_from_slice(&data[start..start + 4]);
-                            let v = i32::from_le_bytes(w);
-                            out.push((v & 0xff) as u8);
-                        }
-                        let any_box: Box<dyn Any> = Box::new(out);
-                        if let Ok(boxed_t) = any_box.downcast::<T>() {
-                            return Ok(Some(*boxed_t));
-                        }
-                    }
-                }
-            }
-
-            match Self::deserialize::<T>(data) {
-                Ok(value) => return Ok(Some(value)),
-                Err(_e) => {
-                    // Memory cache entry is corrupted, remove it and treat as cache miss
-                    self.inner.memory_cache.remove(key);
-                    self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-                    return Ok(None);
-                }
-            }
-        }
-
-        // Try to load from disk
-        let metadata_path = Self::metadata_path(&self.inner, key);
-        let data_path = Self::object_path(&self.inner, key);
-
-        // Acquire read semaphore with timeout to prevent deadlocks
-        let permit =
-            match tokio::time::timeout(Duration::from_secs(5), self.inner.read_semaphore.acquire())
-                .await
-            {
-                Ok(Ok(permit)) => permit,
-                Ok(Err(_)) => {
-                    return Err(CacheError::StoreUnavailable {
-                        store_type: StoreType::Local,
-                        reason: "I/O semaphore closed".to_string(),
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(100),
-                        },
-                    });
-                }
-                Err(_) => {
-                    return Err(CacheError::Timeout {
-                        operation: "acquire read semaphore for get",
-                        duration: Duration::from_secs(5),
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(100),
-                        },
-                    });
-                }
-            };
-
-        // Read metadata first
-        let metadata_bytes = match fs::read(&metadata_path).await {
-            Ok(bytes) => bytes,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                self.inner.stats.misses.fetch_add(1, Ordering::Relaxed);
-                return Ok(None);
-            }
-            Err(e) => {
-                self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-                return Err(CacheError::Io {
-                    path: metadata_path,
-                    operation: "read metadata file",
-                    source: e,
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(100),
-                    },
-                });
-            }
-        };
-
-        let metadata: CacheMetadata = match Self::deserialize(&metadata_bytes) {
-            Ok(m) => m,
-            Err(_e) => {
-                // Release semaphore permit before cleanup to prevent deadlock
-                drop(permit);
-
-                // Metadata is corrupted - clean up both files to prevent future corruption errors
-                let _ = fs::remove_file(&metadata_path).await;
-                let _ = fs::remove_file(&data_path).await;
-                self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-                // Treat as cache miss for better error recovery
-                return Ok(None);
-            }
-        };
-
-        // Check if expired
-        if let Some(expires_at) = metadata.expires_at {
-            if expires_at <= SystemTime::now() {
-                // Release semaphore permit before cleanup to prevent deadlock
-                drop(permit);
-
-                // Remove expired entry
-                let _ = fs::remove_file(&metadata_path).await;
-                let _ = fs::remove_file(&data_path).await;
-                self.inner.stats.misses.fetch_add(1, Ordering::Relaxed);
-                return Ok(None);
-            }
-        }
-
-        // Try to memory-map the data file for zero-copy access
-        let (mmap_option, data) = match Self::mmap_file(&data_path) {
-            Ok(mmap) => {
-                let data = Vec::new(); // Empty vec, we'll use mmap
-                (Some(mmap), data)
-            }
-            Err(_) => {
-                // Fall back to regular file read
-                match fs::read(&data_path).await {
-                    Ok(bytes) => (None, bytes),
-                    Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                        // Release semaphore permit before cleanup to prevent deadlock
-                        drop(permit);
-
-                        // Metadata exists but data doesn't - corrupted state
-                        let _ = fs::remove_file(&metadata_path).await;
-                        self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-                        return Err(CacheError::Corruption {
-                            key: key.to_string(),
-                            reason: "Metadata exists but data is missing".to_string(),
-                            recovery_hint: RecoveryHint::ClearAndRetry,
-                        });
-                    }
-                    Err(e) => {
-                        self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-                        return Err(CacheError::Io {
-                            path: data_path,
-                            operation: "read cache data file",
-                            source: e,
-                            recovery_hint: RecoveryHint::Retry {
-                                after: Duration::from_millis(100),
-                            },
-                        });
-                    }
-                }
-            }
-        };
-
-        // Store in memory cache for hot access — but respect memory limit by skipping insert if it would exceed.
-        let mmap_arc = mmap_option.map(Arc::new);
-        let size = if mmap_arc.is_some() {
-            metadata.size_bytes
-        } else {
-            data.len() as u64
-        };
-
-        let can_store_in_memory = match self.inner.config.max_memory_size {
-            Some(max) => {
-                let current = self.inner.stats.total_bytes.load(Ordering::Relaxed);
-                current.saturating_add(size) <= max
-            }
-            None => true,
-        };
-
-        if can_store_in_memory {
-            let entry = Arc::new(InMemoryEntry {
-                mmap: mmap_arc.clone(),
-                data: data.clone(),
-                metadata: metadata.clone(),
-                last_accessed: RwLock::new(Instant::now()),
-            });
-
-            self.inner
-                .memory_cache
-                .insert(key.to_string(), entry.clone());
-
-            // Record access for eviction policy and update memory stats
-            self.inner.eviction_policy.on_access(key, size);
-            self.inner
-                .stats
-                .total_bytes
-                .fetch_add(size, Ordering::Relaxed);
-        }
-
-        // Count this as a hit regardless of whether we cached in memory
-        self.inner.stats.hits.fetch_add(1, Ordering::Relaxed);
-
-        // Deserialize from the appropriate source
-        let data_slice = if let Some(ref mmap) = mmap_arc {
-            &mmap.as_ref()[..]
-        } else {
-            &data
-        };
-
-        match Self::deserialize::<T>(data_slice) {
-            Ok(value) => Ok(Some(value)),
-            Err(_e) => {
-                // For better error recovery, treat any deserialization error as a cache miss
-                // This is more aggressive than needed but ensures cache remains functional
-
-                // IMPORTANT: We cannot safely do cleanup here while holding memory_cache and other locks
-                // This can cause deadlocks. Instead, just remove from memory cache and return None.
-                // The corrupted disk files will be cleaned up by the background cleanup task.
-
-                // Remove from memory cache as well
-                self.inner.memory_cache.remove(key);
-                self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-                // Return None (cache miss) instead of error for better recovery
-                Ok(None)
-            }
-        }
-    }
-
-    async fn put<T>(&self, key: &str, value: &T, ttl: Option<Duration>) -> Result<()>
-    where
-        T: Serialize + Send + Sync,
-    {
-        match key.validate() {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        let data = match Self::serialize(value) {
-            Ok(d) => d,
-            Err(e) => {
-                // Increment error counter for failed serialization
-                self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-                return Err(e);
-            }
-        };
-
-        // Validate entry size limit using configured max_size_bytes
-        let max_entry_size = self.inner.config.max_size_bytes as usize;
-        if max_entry_size > 0 && data.len() > max_entry_size {
-            self.inner.stats.errors.fetch_add(1, Ordering::Relaxed);
-            return Err(CacheError::CapacityExceeded {
-                requested_bytes: data.len() as u64,
-                available_bytes: max_entry_size as u64,
-                recovery_hint: RecoveryHint::Manual {
-                    instructions: format!(
-                        "Entry size {} bytes exceeds maximum of {} bytes",
-                        data.len(),
-                        max_entry_size
-                    ),
-                },
-            });
-        }
-
-        let now = SystemTime::now();
-
-        // Resolve TTL - use provided TTL or fall back to default TTL
-        let effective_ttl = ttl.or(self.inner.config.default_ttl);
-
-        // Check entry count limit before any insertion
-        let current_entry_count = self.inner.stats.entry_count.load(Ordering::Relaxed);
-        let is_replacing_existing =
-            self.inner.memory_cache.contains_key(key) || self.inner.fast_path.contains_small(key);
-
-        if !is_replacing_existing
-            && self.inner.config.max_entries > 0
-            && current_entry_count >= self.inner.config.max_entries
-        {
-            return Err(CacheError::CapacityExceeded {
-                requested_bytes: data.len() as u64,
-                available_bytes: 0,
-                recovery_hint: RecoveryHint::Manual {
-                    instructions: format!(
-                        "Cache has reached maximum entry limit of {}. Consider increasing max_entries or clearing old entries.",
-                        self.inner.config.max_entries
-                    ),
-                },
-            });
-        }
-
-        // Fast path for small values (< 256 bytes)
-        if data.len() < 256 {
-            let metadata = CacheMetadata {
-                created_at: now,
-                last_accessed: now,
-                expires_at: effective_ttl.map(|d| now + d),
-                size_bytes: data.len() as u64,
-                access_count: 0,
-                content_hash: {
-                    let mut hasher = Sha256::new();
-                    hasher.update(&data);
-                    format!("{:x}", hasher.finalize())
-                },
-                cache_version: self.inner.version,
-            };
-
-            if self
-                .inner
-                .fast_path
-                .put_small(key.to_string(), data.clone(), metadata)
-            {
-                self.inner.stats.writes.fetch_add(1, Ordering::Relaxed);
-                // Only increment entry count if this is a new entry
-                if !is_replacing_existing {
-                    self.inner.stats.entry_count.fetch_add(1, Ordering::Relaxed);
-                }
-                return Ok(());
-            }
-        }
-
-        let metadata = CacheMetadata {
-            created_at: now,
-            last_accessed: now,
-            expires_at: effective_ttl.map(|d| now + d),
-            size_bytes: data.len() as u64,
-            access_count: 0,
-            content_hash: {
-                let mut hasher = Sha256::new();
-                hasher.update(&data);
-                format!("{:x}", hasher.finalize())
-            },
-            cache_version: self.inner.version,
-        };
-
-        // Check if we need to evict due to memory limits before adding new entry
-        if let Some(max_memory) = self.inner.config.max_memory_size {
-            let current_memory = self.inner.stats.total_bytes.load(Ordering::Relaxed);
-            let new_total = current_memory.saturating_add(data.len() as u64);
-
-            if new_total > max_memory {
-                // Run eviction to make space
-                match self.evict_entries().await {
-                    Ok(()) => {}
-                    Err(e) => {
-                        tracing::warn!("Failed to evict entries: {}", e);
-                    }
-                }
-            }
-        }
-
-        // Check memory pressure and disk quota
-        if !self.inner.memory_manager.can_allocate(data.len() as u64) {
-            // Run eviction
-            match self.evict_entries().await {
-                Ok(()) => {}
-                Err(e) => {
-                    tracing::warn!("Failed to evict entries: {}", e);
-                }
-            }
-        }
-
-        match self
-            .inner
-            .memory_manager
-            .check_disk_quota(data.len() as u64)
-        {
-            Ok(_) => {}
-            Err(e) => {
-                // Try eviction first
-                match self.evict_entries().await {
-                    Ok(()) => {
-                        // Retry quota check
-                        match self
-                            .inner
-                            .memory_manager
-                            .check_disk_quota(data.len() as u64)
-                        {
-                            Ok(_) => {}
-                            Err(e) => return Err(e),
-                        }
-                    }
-                    Err(_) => return Err(e),
-                }
-            }
-        }
-
-        // Eviction policy will be notified only if we insert into in-memory cache (see below).
-
-        // Memory limit is enforced for in-memory caching only. If adding this entry would exceed the
-        // configured memory, we skip inserting into the in-memory cache below but still persist to disk.
-        // Eviction is already attempted above to make space; we don't error here to allow disk-only storage.
-        if let Some(_max_memory) = self.inner.config.max_memory_size {
-            // Intentionally no-op: the can_store_in_memory check below governs memory insertion.
-        }
-
-        // Check capacity against max_size_bytes (legacy config)
-        let new_total = self
-            .inner
-            .stats
-            .total_bytes
-            .load(Ordering::Relaxed)
-            .saturating_add(data.len() as u64);
-
-        if self.inner.config.max_size_bytes > 0 && new_total > self.inner.config.max_size_bytes {
-            return Err(CacheError::CapacityExceeded {
-                requested_bytes: data.len() as u64,
-                available_bytes: self
-                    .inner
-                    .config
-                    .max_size_bytes
-                    .saturating_sub(self.inner.stats.total_bytes.load(Ordering::Relaxed)),
-                recovery_hint: RecoveryHint::IncreaseCapacity {
-                    suggested_bytes: new_total,
-                },
-            });
-        }
-
-        // Store in memory for immediate access — but if this would exceed the memory limit,
-        // skip the in-memory insert and rely on disk-only persistence.
-        let can_store_in_memory = match self.inner.config.max_memory_size {
-            Some(max) => {
-                let current = self.inner.stats.total_bytes.load(Ordering::Relaxed);
-                current.saturating_add(data.len() as u64) <= max
-            }
-            None => true,
-        };
-
-        if can_store_in_memory {
-            let entry = Arc::new(InMemoryEntry {
-                mmap: None, // Will be set on next read
-                data: data.clone(),
-                metadata: metadata.clone(),
-                last_accessed: RwLock::new(Instant::now()),
-            });
-
-            // If we're replacing an existing entry, subtract its size
-            let is_new_entry =
-                if let Some(old_entry) = self.inner.memory_cache.insert(key.to_string(), entry) {
-                    self.inner
-                        .stats
-                        .total_bytes
-                        .fetch_sub(old_entry.data.len() as u64, Ordering::Relaxed);
-                    false // Replacing existing entry
-                } else {
-                    true // New entry
-                };
-
-            self.inner
-                .stats
-                .total_bytes
-                .fetch_add(data.len() as u64, Ordering::Relaxed);
-
-            // Update entry count if this is a new entry
-            if is_new_entry {
-                self.inner.stats.entry_count.fetch_add(1, Ordering::Relaxed);
-            }
-
-            // Record insertion with eviction policy only when we added to in-memory cache
-            self.inner.eviction_policy.on_insert(key, data.len() as u64);
-        } else {
-            // Memory is at or above limit. Proactively evict one victim so older entries
-            // are removed from both memory and disk, ensuring recency preference.
-            let mut victim = self.inner.eviction_policy.next_eviction();
-
-            if victim.is_none() {
-                // Fallback: choose the least-recently-accessed in-memory entry.
-                let mut oldest_key: Option<String> = None;
-                let mut oldest_instant = Instant::now();
-                for item in self.inner.memory_cache.iter() {
-                    let last = *item.value().last_accessed.read();
-                    if oldest_key.is_none() || last < oldest_instant {
-                        oldest_instant = last;
-                        oldest_key = Some(item.key().clone());
-                    }
-                }
-                victim = oldest_key;
-            }
-
-            if let Some(v) = victim {
-                // Best-effort eviction; ignore errors to avoid failing this put.
-                let _ = self.remove(&v).await;
-            }
-        }
-
-        // Write to disk
-        let data_path = Self::object_path(&self.inner, key);
-        let data_parent = match data_path.parent() {
-            Some(p) => p,
-            None => {
-                return Err(CacheError::Configuration {
-                    message: "Invalid cache path".to_string(),
-                    recovery_hint: RecoveryHint::Manual {
-                        instructions: "Check cache configuration".to_string(),
-                    },
-                });
-            }
-        };
-
-        // Acquire write semaphore with timeout to prevent deadlocks
-        let _permit = match tokio::time::timeout(
-            Duration::from_secs(5),
-            self.inner.write_semaphore.acquire(),
-        )
-        .await
-        {
-            Ok(Ok(permit)) => permit,
-            Ok(Err(_)) => {
-                return Err(CacheError::StoreUnavailable {
-                    store_type: StoreType::Local,
-                    reason: "Write semaphore closed".to_string(),
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(100),
-                    },
-                });
-            }
-            Err(_) => {
-                return Err(CacheError::Timeout {
-                    operation: "acquire write semaphore for put",
-                    duration: Duration::from_secs(5),
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(100),
-                    },
-                });
-            }
-        };
-
-        match fs::create_dir_all(data_parent).await {
-            Ok(()) => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: data_parent.to_path_buf(),
-                    operation: "create cache directory",
-                    source: e,
-                    recovery_hint: RecoveryHint::CheckPermissions {
-                        path: data_parent.to_path_buf(),
-                    },
-                });
-            }
-        }
-
-        // Serialize metadata separately
-        let metadata_bytes = match Self::serialize(&metadata) {
-            Ok(bytes) => bytes,
-            Err(e) => return Err(e),
-        };
-
-        // Write metadata to separate file for efficient scanning
-        let metadata_path = Self::metadata_path(&self.inner, key);
-        let metadata_parent = match metadata_path.parent() {
-            Some(p) => p,
-            None => {
-                return Err(CacheError::Configuration {
-                    message: "Invalid metadata path".to_string(),
-                    recovery_hint: RecoveryHint::Manual {
-                        instructions: "Check cache configuration".to_string(),
-                    },
-                });
-            }
-        };
-
-        match fs::create_dir_all(metadata_parent).await {
-            Ok(()) => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: metadata_parent.to_path_buf(),
-                    operation: "create metadata directory",
-                    source: e,
-                    recovery_hint: RecoveryHint::CheckPermissions {
-                        path: metadata_parent.to_path_buf(),
-                    },
-                });
-            }
-        }
-
-        // CRITICAL FIX: Write data file FIRST, then metadata
-        // This prevents readers from seeing metadata pointing to incomplete data
-
-        let unique_id = uuid::Uuid::new_v4();
-        let temp_data_path = data_path.with_extension(format!("tmp.{unique_id}"));
-        let temp_metadata_path = metadata_path.with_extension(format!("tmp.{unique_id}"));
-
-        // Step 1: Write data file first
-        match fs::write(&temp_data_path, &data).await {
-            Ok(()) => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: temp_data_path.clone(),
-                    operation: "write cache data file",
-                    source: e,
-                    recovery_hint: RecoveryHint::CheckPermissions {
-                        path: temp_data_path.clone(),
-                    },
-                });
-            }
-        }
-
-        // Step 2: Write metadata file only after data is complete
-        match fs::write(&temp_metadata_path, &metadata_bytes).await {
-            Ok(()) => {}
-            Err(e) => {
-                // Clean up temp data file since metadata write failed
-                let _ = fs::remove_file(&temp_data_path).await;
-                return Err(CacheError::Io {
-                    path: temp_metadata_path.clone(),
-                    operation: "write metadata file",
-                    source: e,
-                    recovery_hint: RecoveryHint::CheckPermissions {
-                        path: temp_metadata_path.clone(),
-                    },
-                });
-            }
-        }
-
-        // Step 3: Rename data file first (so metadata never points to missing data)
-        match fs::rename(&temp_data_path, &data_path).await {
-            Ok(()) => {}
-            Err(e) => {
-                // Clean up temp files
-                let _ = fs::remove_file(&temp_metadata_path).await;
-                let _ = fs::remove_file(&temp_data_path).await;
-                return Err(CacheError::Io {
-                    path: data_path.clone(),
-                    operation: "rename cache data file",
-                    source: e,
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(10),
-                    },
-                });
-            }
-        }
-
-        // Step 4: Finally rename metadata file (data is now available)
-        match fs::rename(&temp_metadata_path, &metadata_path).await {
-            Ok(()) => {}
-            Err(e) => {
-                // Data file exists but metadata rename failed - clean up data file
-                let _ = fs::remove_file(&data_path).await;
-                let _ = fs::remove_file(&temp_metadata_path).await;
-                return Err(CacheError::Io {
-                    path: metadata_path.clone(),
-                    operation: "rename metadata file",
-                    source: e,
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(10),
-                    },
-                });
-            }
-        }
-
-        // Record disk usage for quota tracking
-        self.inner
-            .memory_manager
-            .record_disk_usage(&data_path, data.len() as i64);
-
-        self.inner.stats.writes.fetch_add(1, Ordering::Relaxed);
-        Ok(())
-    }
-
-    async fn remove(&self, key: &str) -> Result<bool> {
-        match key.validate() {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        let mut removed = false;
-
-        // Remove from memory
-        if let Some((_, entry)) = self.inner.memory_cache.remove(key) {
-            let size = if entry.mmap.is_some() {
-                entry.metadata.size_bytes
-            } else {
-                entry.data.len() as u64
-            };
-
-            // Record removal for eviction policy
-            self.inner.eviction_policy.on_remove(key, size);
-
-            self.inner
-                .stats
-                .total_bytes
-                .fetch_sub(size, Ordering::Relaxed);
-            self.inner.stats.entry_count.fetch_sub(1, Ordering::Relaxed);
-            removed = true;
-        }
-
-        // Remove from fast-path cache
-        if self.inner.fast_path.remove_small(key) {
-            removed = true;
-        }
-
-        // Remove from disk (both metadata and data)
-        let metadata_path = Self::metadata_path(&self.inner, key);
-        let data_path = Self::object_path(&self.inner, key);
-
-        match fs::remove_file(&metadata_path).await {
-            Ok(()) => {
-                // Only decrement entry count if it wasn't already removed from memory
-                if !removed {
-                    self.inner.stats.entry_count.fetch_sub(1, Ordering::Relaxed);
-                }
-                removed = true;
-            }
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: metadata_path,
-                    operation: "remove metadata file",
-                    source: e,
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(100),
-                    },
-                });
-            }
-        }
-
-        // Get file size before removal for disk tracking
-        let file_size = match tokio::fs::metadata(&data_path).await {
-            Ok(metadata) => metadata.len() as i64,
-            Err(_) => 0, // File doesn't exist or can't get metadata
-        };
-
-        match fs::remove_file(&data_path).await {
-            Ok(()) => {
-                // Record negative disk usage for removed file
-                if file_size > 0 {
-                    self.inner
-                        .memory_manager
-                        .record_disk_usage(&data_path, -file_size);
-                }
-                removed = true;
-            }
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: data_path,
-                    operation: "remove cache data file",
-                    source: e,
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(100),
-                    },
-                });
-            }
-        }
-
-        if removed {
-            self.inner.stats.removals.fetch_add(1, Ordering::Relaxed);
-        }
-
-        Ok(removed)
-    }
-
-    async fn contains(&self, key: &str) -> Result<bool> {
-        match key.validate() {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        // Check fast-path cache first
-        if self.inner.fast_path.contains_small(key) {
-            return Ok(true);
-        }
-
-        // Check memory cache
-        if let Some(entry) = self.inner.memory_cache.get(key) {
-            // Check if expired
-            if let Some(expires_at) = entry.metadata.expires_at {
-                if expires_at <= SystemTime::now() {
-                    return Ok(false);
-                }
-            }
-            return Ok(true);
-        }
-
-        // Check disk - only need to check metadata
-        let metadata_path = Self::metadata_path(&self.inner, key);
-
-        // Use async metadata check instead of sync exists()
-        let exists = match fs::metadata(&metadata_path).await {
-            Ok(_) => true,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => false,
-            Err(e) => {
-                return Err(CacheError::Io {
-                    path: metadata_path.clone(),
-                    operation: "check metadata file existence",
-                    source: e,
-                    recovery_hint: RecoveryHint::Retry {
-                        after: Duration::from_millis(100),
-                    },
-                });
-            }
-        };
-
-        if !exists {
-            return Ok(false);
-        }
-
-        // Need to check if the disk entry is expired
-        // Acquire read semaphore with timeout to prevent deadlocks
-        let _permit =
-            match tokio::time::timeout(Duration::from_secs(5), self.inner.read_semaphore.acquire())
-                .await
-            {
-                Ok(Ok(permit)) => permit,
-                Ok(Err(_)) => {
-                    return Err(CacheError::StoreUnavailable {
-                        store_type: StoreType::Local,
-                        reason: "Read semaphore closed".to_string(),
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(100),
-                        },
-                    });
-                }
-                Err(_) => {
-                    return Err(CacheError::Timeout {
-                        operation: "acquire read semaphore for contains",
-                        duration: Duration::from_secs(5),
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(100),
-                        },
-                    });
-                }
-            };
-
-        match fs::read(&metadata_path).await {
-            Ok(metadata_bytes) => {
-                let metadata: CacheMetadata = match Self::deserialize(&metadata_bytes) {
-                    Ok(m) => m,
-                    Err(_) => return Ok(false),
-                };
-
-                // Check if expired
-                if let Some(expires_at) = metadata.expires_at {
-                    if expires_at <= SystemTime::now() {
-                        // It's expired, so it doesn't exist
-                        return Ok(false);
-                    }
-                }
-
-                Ok(true)
-            }
-            Err(_) => Ok(false),
-        }
-    }
-
-    async fn metadata(&self, key: &str) -> Result<Option<CacheMetadata>> {
-        match key.validate() {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        // Check fast-path cache first
-        if let Some((_data, metadata)) = self.inner.fast_path.get_small(key) {
-            return Ok(Some(metadata));
-        }
-
-        // Check memory cache
-        if let Some(entry) = self.inner.memory_cache.get(key) {
-            return Ok(Some(entry.metadata.clone()));
-        }
-
-        // Try to load from disk (just metadata)
-        let metadata_path = Self::metadata_path(&self.inner, key);
-        match fs::read(&metadata_path).await {
-            Ok(metadata_bytes) => {
-                let metadata: CacheMetadata = match Self::deserialize(&metadata_bytes) {
-                    Ok(m) => m,
-                    Err(e) => return Err(e),
-                };
-
-                Ok(Some(metadata))
-            }
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(None),
-            Err(e) => Err(CacheError::Io {
-                path: metadata_path,
-                operation: "read metadata file",
-                source: e,
-                recovery_hint: RecoveryHint::Retry {
-                    after: Duration::from_millis(100),
-                },
-            }),
-        }
-    }
-
-    async fn clear(&self) -> Result<()> {
-        // Clear memory cache
-        self.inner.memory_cache.clear();
-        // Clear fast path cache
-        self.inner.fast_path.clear();
-        self.inner.stats.total_bytes.store(0, Ordering::Relaxed);
-        self.inner.stats.entry_count.store(0, Ordering::Relaxed);
-
-        // Clear disk cache
-        let objects_dir = self.inner.base_dir.join("objects");
-        if fs::metadata(&objects_dir).await.is_ok() {
-            match fs::remove_dir_all(&objects_dir).await {
-                Ok(()) => {}
-                Err(e) => {
-                    return Err(CacheError::Io {
-                        path: objects_dir.clone(),
-                        operation: "clear cache directory",
-                        source: e,
-                        recovery_hint: RecoveryHint::CheckPermissions {
-                            path: objects_dir.clone(),
-                        },
-                    });
-                }
-            }
-
-            match fs::create_dir_all(&objects_dir).await {
-                Ok(()) => {}
-                Err(e) => {
-                    return Err(CacheError::Io {
-                        path: objects_dir.clone(),
-                        operation: "recreate cache directory",
-                        source: e,
-                        recovery_hint: RecoveryHint::CheckPermissions { path: objects_dir },
-                    });
-                }
-            }
-        }
-
-        // Also clear metadata directory
-        let metadata_dir = self.inner.base_dir.join("metadata");
-        if fs::metadata(&metadata_dir).await.is_ok() {
-            match fs::remove_dir_all(&metadata_dir).await {
-                Ok(()) => {}
-                Err(e) => {
-                    return Err(CacheError::Io {
-                        path: metadata_dir.clone(),
-                        operation: "clear metadata directory",
-                        source: e,
-                        recovery_hint: RecoveryHint::CheckPermissions {
-                            path: metadata_dir.clone(),
-                        },
-                    });
-                }
-            }
-
-            match fs::create_dir_all(&metadata_dir).await {
-                Ok(()) => {}
-                Err(e) => {
-                    return Err(CacheError::Io {
-                        path: metadata_dir.clone(),
-                        operation: "recreate metadata directory",
-                        source: e,
-                        recovery_hint: RecoveryHint::CheckPermissions { path: metadata_dir },
-                    });
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    async fn statistics(&self) -> Result<CacheStatistics> {
-        let entry_count = self.inner.stats.entry_count.load(Ordering::Relaxed);
-
-        Ok(CacheStatistics {
-            hits: self.inner.stats.hits.load(Ordering::Relaxed),
-            misses: self.inner.stats.misses.load(Ordering::Relaxed),
-            writes: self.inner.stats.writes.load(Ordering::Relaxed),
-            removals: self.inner.stats.removals.load(Ordering::Relaxed),
-            errors: self.inner.stats.errors.load(Ordering::Relaxed),
-            entry_count,
-            total_bytes: self.inner.stats.total_bytes.load(Ordering::Relaxed),
-            max_bytes: self.inner.config.max_size_bytes,
-            expired_cleanups: self.inner.stats.expired_cleanups.load(Ordering::Relaxed),
-            stats_since: self.inner.stats.stats_since,
-            compression_enabled: self.inner.config.compression_enabled,
-            compression_ratio: 1.0, // TODO: Track actual compression ratio
-            wal_recoveries: 0,      // TODO: Track WAL recoveries
-            checksum_failures: 0,   // TODO: Track checksum failures
-        })
-    }
-}
-
-impl fmt::Debug for Cache {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Cache")
-            .field("base_dir", &self.inner.base_dir)
-            .field("version", &self.inner.version)
-            .field("entry_count", &self.inner.memory_cache.len())
-            .finish()
-    }
-}
-
-impl Drop for CacheInner {
-    fn drop(&mut self) {
-        // Cancel cleanup task
-        // parking_lot RwLock doesn't block on drop, so this is safe
-        if let Some(handle) = self.cleanup_handle.write().take() {
-            handle.abort();
-        }
-    }
-}
-
-impl StreamingCache for Cache {
-    fn get_reader<'a>(
-        &'a self,
-        key: &'a str,
-    ) -> Pin<Box<dyn Future<Output = Result<Option<CacheReader>>> + Send + 'a>> {
-        Box::pin(async move {
-            match key.validate() {
-                Ok(()) => {}
-                Err(e) => return Err(e),
-            }
-
-            // Check if the entry exists and get metadata
-            let metadata_path = Self::metadata_path(&self.inner, key);
-            let data_path = Self::object_path(&self.inner, key);
-
-            // Acquire read semaphore with timeout to prevent deadlocks
-            let _permit = match tokio::time::timeout(
-                Duration::from_secs(5),
-                self.inner.read_semaphore.acquire(),
-            )
-            .await
-            {
-                Ok(Ok(permit)) => permit,
-                Ok(Err(_)) => {
-                    return Err(CacheError::StoreUnavailable {
-                        store_type: StoreType::Local,
-                        reason: "Read semaphore closed".to_string(),
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(100),
-                        },
-                    });
-                }
-                Err(_) => {
-                    return Err(CacheError::Timeout {
-                        operation: "acquire read semaphore for streaming",
-                        duration: Duration::from_secs(5),
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(100),
-                        },
-                    });
-                }
-            };
-
-            // Read metadata
-            let metadata_bytes = match fs::read(&metadata_path).await {
-                Ok(bytes) => bytes,
-                Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                    return Ok(None);
-                }
-                Err(e) => {
-                    return Err(CacheError::Io {
-                        path: metadata_path,
-                        operation: "read metadata for streaming",
-                        source: e,
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(100),
-                        },
-                    });
-                }
-            };
-
-            let metadata: CacheMetadata = match Self::deserialize(&metadata_bytes) {
-                Ok(m) => m,
-                Err(e) => return Err(e),
-            };
-
-            // Check if expired
-            if let Some(expires_at) = metadata.expires_at {
-                if expires_at <= SystemTime::now() {
-                    // Remove expired entry
-                    let _ = fs::remove_file(&metadata_path).await;
-                    let _ = fs::remove_file(&data_path).await;
-                    return Ok(None);
-                }
-            }
-
-            // Update access time in memory cache if present
-            if let Some(entry) = self.inner.memory_cache.get(key) {
-                *entry.last_accessed.write() = Instant::now();
-            }
-
-            // Prefer memory-mapped reader for performance
-            #[cfg(target_os = "linux")]
-            {
-                match CacheReader::from_mmap(data_path.clone(), metadata.clone()).await {
-                    Ok(reader) => return Ok(Some(reader)),
-                    Err(_) => {
-                        // Fall back to regular file reader
-                    }
-                }
-            }
-
-            // Use file-based reader
-            match CacheReader::from_file(data_path, metadata).await {
-                Ok(reader) => Ok(Some(reader)),
-                Err(e) => Err(e),
-            }
-        })
-    }
-
-    fn get_writer<'a>(
-        &'a self,
-        key: &'a str,
-        ttl: Option<Duration>,
-    ) -> Pin<Box<dyn Future<Output = Result<CacheWriter>> + Send + 'a>> {
-        Box::pin(async move {
-            match key.validate() {
-                Ok(()) => {}
-                Err(e) => return Err(e),
-            }
-
-            // Check capacity before creating writer
-            if self.inner.config.max_size_bytes > 0 {
-                let current_size = self.inner.stats.total_bytes.load(Ordering::Relaxed);
-                if current_size >= self.inner.config.max_size_bytes {
-                    return Err(CacheError::CapacityExceeded {
-                        requested_bytes: 0,
-                        available_bytes: 0,
-                        recovery_hint: RecoveryHint::IncreaseCapacity {
-                            suggested_bytes: self.inner.config.max_size_bytes * 2,
-                        },
-                    });
-                }
-            }
-
-            match CacheWriter::new(&self.inner.base_dir, key, ttl).await {
-                Ok(writer) => Ok(writer),
-                Err(e) => Err(e),
-            }
-        })
-    }
-
-    fn put_stream<'a, R>(
-        &'a self,
-        key: &'a str,
-        reader: R,
-        ttl: Option<Duration>,
-    ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'a>>
-    where
-        R: AsyncRead + Send + 'a,
-    {
-        Box::pin(async move {
-            let mut writer = match self.get_writer(key, ttl).await {
-                Ok(w) => w,
-                Err(e) => return Err(e),
-            };
-
-            // High-performance streaming copy
-            const BUFFER_SIZE: usize = 64 * 1024; // 64KB buffer
-            let mut buffer = vec![0u8; BUFFER_SIZE];
-            let mut total_bytes = 0u64;
-
-            tokio::pin!(reader);
-
-            loop {
-                let n = match reader.read(&mut buffer).await {
-                    Ok(0) => break, // EOF
-                    Ok(n) => n,
-                    Err(e) => {
-                        return Err(CacheError::Io {
-                            path: PathBuf::from(key),
-                            operation: "read from stream",
-                            source: std::io::Error::other(e),
-                            recovery_hint: RecoveryHint::Retry {
-                                after: Duration::from_millis(100),
-                            },
-                        });
-                    }
-                };
-
-                match writer.write_all(&buffer[..n]).await {
-                    Ok(()) => {}
-                    Err(e) => {
-                        return Err(CacheError::Io {
-                            path: PathBuf::from(key),
-                            operation: "write to cache stream",
-                            source: std::io::Error::other(e),
-                            recovery_hint: RecoveryHint::Retry {
-                                after: Duration::from_millis(100),
-                            },
-                        });
-                    }
-                }
-
-                total_bytes += n as u64;
-            }
-
-            // Finalize the write
-            let metadata = match writer.finalize().await {
-                Ok(m) => m,
-                Err(e) => return Err(e),
-            };
-
-            // Update statistics
-            self.inner.stats.writes.fetch_add(1, Ordering::Relaxed);
-            self.inner
-                .stats
-                .total_bytes
-                .fetch_add(total_bytes, Ordering::Relaxed);
-
-            // Add to memory cache for hot access
-            let _hash = Self::hash_key(&self.inner, key);
-            let data_path = Self::object_path(&self.inner, key);
-
-            // Try to memory-map for future reads
-            let mmap_option = Self::mmap_file(&data_path).ok().map(Arc::new);
-
-            let entry = Arc::new(InMemoryEntry {
-                mmap: mmap_option,
-                data: Vec::new(), // Empty for streamed entries
-                metadata,
-                last_accessed: RwLock::new(Instant::now()),
-            });
-
-            self.inner.memory_cache.insert(key.to_string(), entry);
-
-            Ok(total_bytes)
-        })
-    }
-
-    fn get_stream<'a, W>(
-        &'a self,
-        key: &'a str,
-        writer: W,
-    ) -> Pin<Box<dyn Future<Output = Result<Option<u64>>> + Send + 'a>>
-    where
-        W: AsyncWrite + Send + 'a,
-    {
-        Box::pin(async move {
-            let reader = match self.get_reader(key).await {
-                Ok(Some(r)) => r,
-                Ok(None) => return Ok(None),
-                Err(e) => return Err(e),
-            };
-
-            let _expected_size = reader.metadata().size_bytes;
-
-            // High-performance streaming copy
-            // Note: Zero-copy implementation would be added here for Linux systems
-            // using sendfile/splice system calls for optimal performance
-
-            // Standard async copy
-            let mut reader = reader;
-            const BUFFER_SIZE: usize = 64 * 1024; // 64KB buffer
-            let mut buffer = vec![0u8; BUFFER_SIZE];
-            let mut total_bytes = 0u64;
-
-            tokio::pin!(writer);
-
-            loop {
-                let n = match reader.read(&mut buffer).await {
-                    Ok(0) => break, // EOF
-                    Ok(n) => n,
-                    Err(e) => {
-                        return Err(CacheError::Io {
-                            path: PathBuf::from(key),
-                            operation: "read from cache stream",
-                            source: std::io::Error::other(e),
-                            recovery_hint: RecoveryHint::Retry {
-                                after: Duration::from_millis(100),
-                            },
-                        });
-                    }
-                };
-
-                match writer.write_all(&buffer[..n]).await {
-                    Ok(()) => {}
-                    Err(e) => {
-                        return Err(CacheError::Io {
-                            path: PathBuf::from(key),
-                            operation: "write to output stream",
-                            source: std::io::Error::other(e),
-                            recovery_hint: RecoveryHint::Retry {
-                                after: Duration::from_millis(100),
-                            },
-                        });
-                    }
-                }
-
-                total_bytes += n as u64;
-            }
-
-            match writer.flush().await {
-                Ok(()) => {}
-                Err(e) => {
-                    return Err(CacheError::Io {
-                        path: PathBuf::from(key),
-                        operation: "flush output stream",
-                        source: std::io::Error::other(e),
-                        recovery_hint: RecoveryHint::Retry {
-                            after: Duration::from_millis(10),
-                        },
-                    });
-                }
-            }
-
-            self.inner.stats.hits.fetch_add(1, Ordering::Relaxed);
-            Ok(Some(total_bytes))
-        })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[tokio::test]
-    async fn test_basic_operations() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let cache = Cache::new(temp_dir.path().to_path_buf(), CacheConfig::default()).await?;
-
-        // Test put and get
-        match cache.put("key1", &"value1", None).await {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        let value: Option<String> = match cache.get("key1").await {
-            Ok(v) => v,
-            Err(e) => return Err(e),
-        };
-        assert_eq!(value, Some("value1".to_string()));
-
-        // Test contains
-        match cache.contains("key1").await {
-            Ok(true) => {}
-            Ok(false) => panic!("Key should exist"),
-            Err(e) => return Err(e),
-        }
-
-        match cache.contains("key2").await {
-            Ok(false) => {}
-            Ok(true) => panic!("Key should not exist"),
-            Err(e) => return Err(e),
-        }
-
-        // Test remove
-        match cache.remove("key1").await {
-            Ok(true) => {}
-            Ok(false) => panic!("Key should have been removed"),
-            Err(e) => return Err(e),
-        }
-
-        match cache.contains("key1").await {
-            Ok(false) => {}
-            Ok(true) => panic!("Key should not exist after removal"),
-            Err(e) => return Err(e),
-        }
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_expiration() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let cache = Cache::new(temp_dir.path().to_path_buf(), CacheConfig::default()).await?;
-
-        // Put with short TTL
-        match cache
-            .put("expires", &"soon", Some(Duration::from_millis(50)))
-            .await
-        {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        // Should exist immediately
-        match cache.contains("expires").await {
-            Ok(true) => {}
-            Ok(false) => panic!("Key should exist"),
-            Err(e) => return Err(e),
-        }
-
-        // Wait for expiration
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // Should be expired
-        let value: Option<String> = match cache.get("expires").await {
-            Ok(v) => v,
-            Err(e) => return Err(e),
-        };
-        assert_eq!(value, None);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_entry_limit_enforcement() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let mut config = CacheConfig::default();
-        config.max_entries = 5; // Set small limit for testing
-        let cache = Cache::new(temp_dir.path().to_path_buf(), config).await?;
-
-        // Add entries up to the limit
-        for i in 0..5 {
-            match cache
-                .put(&format!("key_{i}"), &format!("value_{i}"), None)
-                .await
-            {
-                Ok(()) => {}
-                Err(e) => return Err(e),
-            }
-        }
-
-        let stats = cache.statistics().await?;
-        assert_eq!(stats.entry_count, 5);
-
-        // Try to add one more entry - should fail
-        match cache.put("key_6", &"value_6", None).await {
-            Ok(()) => panic!("Should have failed due to entry limit"),
-            Err(CacheError::CapacityExceeded { .. }) => {
-                // Expected
-            }
-            Err(e) => return Err(e),
-        }
-
-        // Statistics should still show 5 entries
-        let stats = cache.statistics().await?;
-        assert_eq!(stats.entry_count, 5);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_statistics() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let cache = Cache::new(temp_dir.path().to_path_buf(), CacheConfig::default()).await?;
-
-        match cache.put("key1", &"value1", None).await {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        let _: Option<String> = match cache.get("key1").await {
-            Ok(v) => v,
-            Err(e) => return Err(e),
-        }; // Hit
-
-        let _: Option<String> = match cache.get("key2").await {
-            Ok(v) => v,
-            Err(e) => return Err(e),
-        }; // Miss
-
-        match cache.remove("key1").await {
-            Ok(_) => {}
-            Err(e) => return Err(e),
-        }
-
-        let stats = match cache.statistics().await {
-            Ok(s) => s,
-            Err(e) => return Err(e),
-        };
-
-        assert_eq!(stats.writes, 1);
-        assert_eq!(stats.hits, 1);
-        assert_eq!(stats.misses, 1);
-        assert_eq!(stats.removals, 1);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_zero_copy_mmap() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let cache = Cache::new(temp_dir.path().to_path_buf(), CacheConfig::default()).await?;
-
-        // Write a large value (but under the entry size limit)
-        let large_data = vec![0u8; 8192]; // 8KB
-        match cache.put("large", &large_data, None).await {
-            Ok(()) => {}
-            Err(e) => return Err(e),
-        }
-
-        // Clear memory cache to force disk read
-        cache.inner.memory_cache.clear();
-
-        // Read should use mmap
-        let value: Option<Vec<u8>> = match cache.get("large").await {
-            Ok(v) => v,
-            Err(e) => return Err(e),
-        };
-
-        assert_eq!(value, Some(large_data));
-
-        // Check that entry in memory cache has mmap
-        if let Some(entry) = cache.inner.memory_cache.get("large") {
-            assert!(entry.mmap.is_some(), "Should have memory-mapped the file");
-        }
-
-        Ok(())
-    }
-
-    // Property test removed to prevent hanging and resource exhaustion
-    // The functionality is adequately covered by the unit tests above
-    #[tokio::test]
-    async fn test_cache_consistency_simple() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let cache = Cache::new(temp_dir.path().to_path_buf(), CacheConfig::default()).await?;
-
-        let test_cases = vec![
-            ("key1".to_string(), "value1".to_string()),
-            ("key2".to_string(), "value2".to_string()),
-            ("key3".to_string(), "value3".to_string()),
-        ];
-
-        // Put all key-value pairs
-        for (key, value) in &test_cases {
-            cache.put(key, value, None).await?;
-        }
-
-        // Verify all can be retrieved
-        for (key, expected_value) in &test_cases {
-            let actual: Option<String> = cache.get(key).await?;
-            assert_eq!(actual.as_ref(), Some(expected_value));
-        }
-
-        // Clear cache
-        cache.clear().await?;
-
-        // Verify all are gone
-        for (key, _) in &test_cases {
-            let value: Option<String> = cache.get(key).await?;
-            assert_eq!(value, None);
-        }
-
-        Ok(())
-    }
-}
