+++ crates/cache/src/core/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Tests for the unified cache implementation
+
+#[cfg(test)]
+mod advanced;
+#[cfg(test)]
+mod basic;
